mcp-database-webui-main/mcp-database-webui-main/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
mcp_use
mysql-mcp-server
langchain_anthropic